// src/payload/payload-types.ts
export interface CollectionWithRichTextField {
  id: string;
  richTextField: {
    root: {
      type: string;
      children: {
        type: string;
        version: number;
        [k: string]: unknown;
      }[];
      direction: ('ltr' | 'rtl') | null;
      format: 'left' | 'start' | 'center' | 'right' | 'end' | 'justify' | '';
      indent: number;
      version: number;
    };
    [k: string]: unknown;
  };
  updatedAt: string;
  createdAt: string;
}
// ...
export interface DocxSetting {
  id: string;
  header?: {
    root: {
      type: string;
      children: {
        type: string;
        version: number;
        [k: string]: unknown;
      }[];
      direction: ('ltr' | 'rtl') | null;
      format: 'left' | 'start' | 'center' | 'right' | 'end' | 'justify' | '';
      indent: number;
      version: number;
    };
    [k: string]: unknown;
  } | null;
  footer?: {
    root: {
      type: string;
      children: {
        type: string;
        version: number;
        [k: string]: unknown;
      }[];
      direction: ('ltr' | 'rtl') | null;
      format: 'left' | 'start' | 'center' | 'right' | 'end' | 'justify' | '';
      indent: number;
      version: number;
    };
    [k: string]: unknown;
  } | null;
  /**
   * Portrait or Landscape
   */
  orientation?: ('portrait' | 'landscape') | null;
  /**
   * A4, Letter, etc.
   */
  pageSize?: ('A4' | 'Letter' | 'Legal') | null;
  margins?: {
    top?: number | null;
    bottom?: number | null;
    left?: number | null;
    right?: number | null;
  };
  style?: {
    fontFamily?: string | null;
    fontSize?: number | null;
    textColor?: string | null;
    /**
     * 1 = single, 1.15 = default, 1.5 = 1.5, etc.
     */
    lineSpacing?: number | null;
  };
  updatedAt?: string | null;
  createdAt?: string | null;
}